successful backup and restore
This is a working PoC. The only thing that isn't working is restore over
a network; local restore works, encryption and decryption works,
filenames work.

One thing we need is some way to communicate path and file size to the
server before the file begins downloading. We might want to save this
with the file, like on the end of it, for scanning later.